process: Fix mask size in ssig
We don't need a 32bit value any more, and using 16bits saves us 100 bytes
(That's how expensive SDCC longs are!)

+++ Kernel/process.c
@@ -593,7 +593,7 @@ rescan:
 void ssig(ptptr proc, uint8_t sig)
 {
 	struct sigbits *m = proc->p_sig;
-	uint32_t sigm;
+	uint16_t sigm;
 	uint8_t sigbit = sig & 0x0F;
 	irqflags_t irq;
